[Adventure]
Load bundled data on cog load and store data locally for the cog. Restructured bundled data to allow for custom themes in the future.

Renamed Dragons variable to Bosses for future theme use. Planned to replace use of Dragon and Basilisk in messages with variables in bundled data.

Add loadouts to save current set of gear and then re-apply that set later. If the item no longer exists in the players inventory and not equipped it's removed from the loadout.

Fix bug with Basilisks.

Fix Bug where sometimes the bot can't find member.id when playing a game.

Fix docstrings for cart setup.

Cleanup stats display to look similar to backpacks display.

Add simple instruction manual to README